Add compile namespace and folder
